substrate/frame/membership/src/tests.rs
// This file is part of Substrate.

// Copyright (C) Parity Technologies (UK) Ltd.
// SPDX-License-Identifier: Apache-2.0

// Licensed under the Apache License, Version 2.0 (the "License");
// you may not use this file except in compliance with the License.
// You may obtain a copy of the License at
//
// 	http://www.apache.org/licenses/LICENSE-2.0
//
// Unless required by applicable law or agreed to in writing, software
// distributed under the License is distributed on an "AS IS" BASIS,
// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
// See the License for the specific language governing permissions and
// limitations under the License.

//! Tests for the module.

use crate as pallet_membership;
use crate::{mock::*, Error};

use sp_runtime::{bounded_vec, traits::BadOrigin, BuildStorage};

use frame_support::{
    assert_noop, assert_ok, assert_storage_noop, 
    traits::{StorageVersion},
};
use frame_support::traits::PalletInfo;

#[test]
	fn query_membership_works() {
    new_test_ext().execute_with(|| {
        assert_eq!(Membership::members(), vec![10, 20, 30]);
        assert_eq!(MEMBERS.with(|m| m.borrow().clone()), vec![10, 20, 30]);
    });
}

#[test]
fn prime_member_works() {
    new_test_ext().execute_with(|| {
        assert_noop!(Membership::set_prime(RuntimeOrigin::signed(4), 20), BadOrigin);
        assert_noop!(
            Membership::set_prime(RuntimeOrigin::signed(5), 15),
            Error::<Test, _>::NotMember
        );
        assert_ok!(Membership::set_prime(RuntimeOrigin::signed(5), 20));
        assert_eq!(Membership::prime(), Some(20));
        assert_eq!(PRIME.with(|m| *m.borrow()), Membership::prime());

        assert_ok!(Membership::clear_prime(RuntimeOrigin::signed(5)));
        assert_eq!(Membership::prime(), None);
        assert_eq!(PRIME.with(|m| *m.borrow()), Membership::prime());
    });
}

#[test]
fn add_member_works() {
    new_test_ext().execute_with(|| {
        assert_noop!(Membership::add_member(RuntimeOrigin::signed(5), 15), BadOrigin);
        assert_noop!(
            Membership::add_member(RuntimeOrigin::signed(1), 10),
            Error::<Test, _>::AlreadyMember
        );
        assert_ok!(Membership::add_member(RuntimeOrigin::signed(1), 15));
        assert_eq!(Membership::members(), vec![10, 15, 20, 30]);
        assert_eq!(MEMBERS.with(|m| m.borrow().clone()), Membership::members().to_vec());
    });
}

#[test]
fn remove_member_works() {
    new_test_ext().execute_with(|| {
        assert_noop!(Membership::remove_member(RuntimeOrigin::signed(5), 20), BadOrigin);
        assert_noop!(
            Membership::remove_member(RuntimeOrigin::signed(2), 15),
            Error::<Test, _>::NotMember
        );
        assert_ok!(Membership::set_prime(RuntimeOrigin::signed(5), 20));
        assert_ok!(Membership::remove_member(RuntimeOrigin::signed(2), 20));
        assert_eq!(Membership::members(), vec![10, 30]);
        assert_eq!(MEMBERS.with(|m| m.borrow().clone()), Membership::members().to_vec());
        assert_eq!(Membership::prime(), None);
        assert_eq!(PRIME.with(|m| *m.borrow()), Membership::prime());
    });
}

#[test]
fn swap_member_works() {
    new_test_ext().execute_with(|| {
        assert_noop!(Membership::swap_member(RuntimeOrigin::signed(5), 10, 25), BadOrigin);
        assert_noop!(
            Membership::swap_member(RuntimeOrigin::signed(3), 15, 25),
            Error::<Test, _>::NotMember
        );
        assert_noop!(
            Membership::swap_member(RuntimeOrigin::signed(3), 10, 30),
            Error::<Test, _>::AlreadyMember
        );

        assert_ok!(Membership::set_prime(RuntimeOrigin::signed(5), 20));
        assert_ok!(Membership::swap_member(RuntimeOrigin::signed(3), 20, 20));
        assert_eq!(Membership::members(), vec![10, 20, 30]);
        assert_eq!(Membership::prime(), Some(20));
        assert_eq!(PRIME.with(|m| *m.borrow()), Membership::prime());

        assert_ok!(Membership::set_prime(RuntimeOrigin::signed(5), 10));
        assert_ok!(Membership::swap_member(RuntimeOrigin::signed(3), 10, 25));
        assert_eq!(Membership::members(), vec![20, 25, 30]);
        assert_eq!(MEMBERS.with(|m| m.borrow().clone()), Membership::members().to_vec());
        assert_eq!(Membership::prime(), None);
        assert_eq!(PRIME.with(|m| *m.borrow()), Membership::prime());
    });
}

#[test]
fn swap_member_works_that_does_not_change_order() {
    new_test_ext().execute_with(|| {
        assert_ok!(Membership::swap_member(RuntimeOrigin::signed(3), 10, 5));
        assert_eq!(Membership::members(), vec![5, 20, 30]);
        assert_eq!(MEMBERS.with(|m| m.borrow().clone()), Membership::members().to_vec());
    });
}

#[test]
fn swap_member_with_identical_arguments_changes_nothing() {
    new_test_ext().execute_with(|| {
        assert_storage_noop!(assert_ok!(Membership::swap_member(
            RuntimeOrigin::signed(3),
            10,
            10
        )));
    });
}

#[test]
fn change_key_works() {
    new_test_ext().execute_with(|| {
        assert_ok!(Membership::set_prime(RuntimeOrigin::signed(5), 10));
        assert_noop!(
            Membership::change_key(RuntimeOrigin::signed(3), 25),
            Error::<Test, _>::NotMember
        );
        assert_noop!(
            Membership::change_key(RuntimeOrigin::signed(10), 20),
            Error::<Test, _>::AlreadyMember
        );
        assert_ok!(Membership::change_key(RuntimeOrigin::signed(10), 40));
        assert_eq!(Membership::members(), vec![20, 30, 40]);
        assert_eq!(MEMBERS.with(|m| m.borrow().clone()), Membership::members().to_vec());
        assert_eq!(Membership::prime(), Some(40));
        assert_eq!(PRIME.with(|m| *m.borrow()), Membership::prime());
    });
}

#[test]
fn change_key_works_that_does_not_change_order() {
    new_test_ext().execute_with(|| {
        assert_ok!(Membership::change_key(RuntimeOrigin::signed(10), 5));
        assert_eq!(Membership::members(), vec![5, 20, 30]);
        assert_eq!(MEMBERS.with(|m| m.borrow().clone()), Membership::members().to_vec());
    });
}

#[test]
fn change_key_with_same_caller_as_argument_changes_nothing() {
    new_test_ext().execute_with(|| {
        assert_storage_noop!(assert_ok!(Membership::change_key(RuntimeOrigin::signed(10), 10)));
    });
}

#[test]
fn reset_members_works() {
    new_test_ext().execute_with(|| {
        assert_ok!(Membership::set_prime(RuntimeOrigin::signed(5), 20));
        assert_noop!(
            Membership::reset_members(RuntimeOrigin::signed(1), bounded_vec![20, 40, 30]),
            BadOrigin
        );

        assert_ok!(Membership::reset_members(RuntimeOrigin::signed(4), vec![20, 40, 30]));
        assert_eq!(Membership::members(), vec![20, 30, 40]);
        assert_eq!(MEMBERS.with(|m| m.borrow().clone()), Membership::members().to_vec());
        assert_eq!(Membership::prime(), Some(20));
        assert_eq!(PRIME.with(|m| *m.borrow()), Membership::prime());

        assert_ok!(Membership::reset_members(RuntimeOrigin::signed(4), vec![10, 40, 30]));
        assert_eq!(Membership::members(), vec![10, 30, 40]);
        assert_eq!(MEMBERS.with(|m| m.borrow().clone()), Membership::members().to_vec());
        assert_eq!(Membership::prime(), None);
        assert_eq!(PRIME.with(|m| *m.borrow()), Membership::prime());
    });
}

#[test]
#[should_panic(expected = "Members cannot contain duplicate accounts.")]
fn genesis_build_panics_with_duplicate_members() {
    pallet_membership::GenesisConfig::<Test> {
        members: bounded_vec![1, 2, 3, 1],
        phantom: Default::default(),
    }
    .build_storage()
    .unwrap();
}

#[test]
fn migration_v4() {
    new_test_ext().execute_with(|| {
        //use frame_support::traits::PalletInfo;
        let old_pallet_name = "OldMembership";
        let new_pallet_name =
            <Test as frame_system::Config>::PalletInfo::name::<Membership>().unwrap();

        frame_support::storage::migration::move_pallet(
            new_pallet_name.as_bytes(),
            old_pallet_name.as_bytes(),
        );

        StorageVersion::new(0).put::<Membership>();

        crate::migrations::v4::pre_migrate::<Membership, _>(old_pallet_name, new_pallet_name);
        crate::migrations::v4::migrate::<Test, Membership, _>(old_pallet_name, new_pallet_name);
        crate::migrations::v4::post_migrate::<Membership, _>(old_pallet_name, new_pallet_name);
    });
}
